hait_advanced/requirements.txt
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-registration==3.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
